+++ src/hooks/useUsers.test.ts
@@ -3,10 +3,8 @@ import { renderHook, waitFor } from '@testing-library/react'
 import { useUsersQuery, useCreateUser, useUpdateUser, useDeleteUser } from '@/hooks/useUsers'
 import { server } from '@/test/server'
 import { http, HttpResponse } from 'msw'
-import { TEST_API_BASE as BASE } from '@/test/constants'
+import { TEST_API_BASE as BASE, TEST_DOMAIN_ID as DOMAIN_ID } from '@/test/constants'
 import { makeQueryWrapper } from '@/test/makeQueryWrapper'
-
-const DOMAIN_ID = 'dom-1'
 
 describe('useUsersQuery', () => {
   let wrapper: ReturnType<typeof makeQueryWrapper>['wrapper']